factorize parents
